+++ copasi/UI/CQFunctionDM.cpp
@@ -46,7 +46,7 @@ size_t CQFunctionDM::size() const
 
 int CQFunctionDM::rowCount(const QModelIndex& C_UNUSED(parent)) const
 {
-  return mFetched + 1;
+  return (int)mFetched + 1;
 }
 
 int CQFunctionDM::columnCount(const QModelIndex& C_UNUSED(parent)) const
@@ -289,7 +289,7 @@ bool CQFunctionDM::removeRows(int position, int rows, const QModelIndex & parent
         }
     }
 
-  beginRemoveRows(parent, position, std::min< int >(mFetched, position + rows) - 1);
+  beginRemoveRows(parent, position, std::min< int >((int)mFetched, position + rows) - 1);
 
   for (itDeletedKey = DeletedKeys.begin(), itDeletedCN = DeletedCNs.begin(), row = 0; itDeletedKey != endDeletedKey; ++itDeletedKey, ++itDeletedCN, ++row)
     {
