+++ copasi/UI/CQLayoutsWidget.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2024 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -168,7 +168,6 @@ void CQLayoutsWidget::deleteSelectedLayouts()
   mpLayoutsDM->removeRows(mappedSelRows);
 }
 
-
 // virtual
 bool CQLayoutsWidget::enterProtected()
 {
@@ -277,7 +276,7 @@ void CQLayoutsWidget::slotBtnNewClicked()
   // update the table
   mpLayoutsDM->insertRows(mpLayoutsDM->rowCount(), 1);
   dataChanged(QModelIndex(), QModelIndex());
-  LayoutWindow *window = createLayoutWindow(pListOfLayouts->size() - 1, pLayout);
+  LayoutWindow *window = createLayoutWindow((int)pListOfLayouts->size() - 1, pLayout);
 #ifndef DISABLE_QT_LAYOUT_RENDERING
   CQAnimationWindow *pAnim = dynamic_cast<CQAnimationWindow *>(window);
 
@@ -328,7 +327,6 @@ void CQLayoutsWidget::slotBtnClearClicked()
       deleteLayoutWindows();
     }
 }
-
 
 // virtual
 void CQLayoutsWidget::slotDoubleClicked(const QModelIndex proxyIndex)
