+++ copasi/UI/CQTrajectoryWidget.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2024 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -160,7 +160,7 @@ void CQTrajectoryWidget::slotIntervals()
 
   try
     {
-      mpTrajectoryProblem->setStepNumber(mpEditIntervals->text().toULong());
+      mpTrajectoryProblem->setStepNumber(mpEditIntervals->text().toUInt());
     }
   catch (...)
     {
